% Auto-generated: do not edit by hand
\name{''DashDraggableResponsive}

\alias{''DashDraggableResponsive}

\title{DashDraggableResponsive component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''DashDraggableResponsive(children=NULL, id=NULL, value=NULL, layout=NULL, ncols=NULL,
nrows=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers | a list of or a singular dash component, string or number. }

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{value}{Character. The value displayed in the input.}

\item{layout}{List of named lists. }

\item{ncols}{Numeric. }

\item{nrows}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

